+++ pkg/client/rules_test.go
@@ -2,7 +2,6 @@ package client
 
 import (
 	"context"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
@@ -10,12 +9,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestCortexClient_X(t *testing.T) {
+func TestCortexClient_DeleteRuleGroup(t *testing.T) {
 	requestCh := make(chan *http.Request, 1)
 
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		requestCh <- r
-		fmt.Fprintln(w, "hello")
 	}))
 	defer ts.Close()
 
@@ -57,7 +55,7 @@ func TestCortexClient_X(t *testing.T) {
 			expURLPath: "/api/v1/rules/My%2FNamespace/%2Ffirst-char-slash",
 		},
 		{
-			test:       "special-characters-slash-first",
+			test:       "special-characters-slash-last",
 			namespace:  "My/Namespace",
 			name:       "last-char-slash/",
 			expURLPath: "/api/v1/rules/My%2FNamespace/last-char-slash%2F",
@@ -69,7 +67,65 @@ func TestCortexClient_X(t *testing.T) {
 
 			req := <-requestCh
 			require.Equal(t, tc.expURLPath, req.URL.EscapedPath())
+			require.Equal(t, http.MethodDelete, req.Method)
+		})
+	}
 
+}
+
+func TestCortexClient_DeleteRuleNamespace(t *testing.T) {
+	requestCh := make(chan *http.Request, 1)
+
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		requestCh <- r
+	}))
+	defer ts.Close()
+
+	client, err := New(Config{
+		Address: ts.URL,
+		ID:      "my-id",
+		Key:     "my-key",
+	})
+	require.NoError(t, err)
+
+	for _, tc := range []struct {
+		test       string
+		namespace  string
+		expURLPath string
+	}{
+		{
+			test:       "regular-characters",
+			namespace:  "my-namespace",
+			expURLPath: "/api/v1/rules/my-namespace",
+		},
+		{
+			test:       "special-characters-spaces",
+			namespace:  "My: Namespace",
+			expURLPath: "/api/v1/rules/My:%20Namespace",
+		},
+		{
+			test:       "special-characters-slashes",
+			namespace:  "My/Namespace",
+			expURLPath: "/api/v1/rules/My%2FNamespace",
+		},
+		{
+			test:       "special-characters-slash-first",
+			namespace:  "/My/Namespace",
+			expURLPath: "/api/v1/rules/%2FMy%2FNamespace",
+		},
+		{
+			test:       "special-characters-slash-last",
+			namespace:  "My/Namespace/",
+			expURLPath: "/api/v1/rules/My%2FNamespace%2F",
+		},
+	} {
+		t.Run(tc.test, func(t *testing.T) {
+			ctx := context.Background()
+			require.NoError(t, client.DeleteRuleNamespace(ctx, tc.namespace))
+
+			req := <-requestCh
+			require.Equal(t, tc.expURLPath, req.URL.EscapedPath())
+			require.Equal(t, http.MethodDelete, req.Method)
 		})
 	}